Fixed two more @*#%! 32-bit errors
Apparently these only show up if you build for "Generic iOS Device",
not a real device. Which is why I didn't encounter them.

+++ Objective-C/CBLArray.mm
@@ -48,7 +48,8 @@
 
 
 [[noreturn]] static void throwRangeException(NSUInteger index) {
-    [NSException raise: NSRangeException format: @"CBLArray index %zu is out of range", index];
+    [NSException raise: NSRangeException format: @"CBLArray index %lu is out of range",
+        (unsigned long)index];
     abort();
 }
